requirements.txt
spacy-udpipe
python-dotenv
langdetect
bitermplus
pandas
numpy
gensim
pyyaml
dicttoxml
nltk